make substitution_matrix_sites faster for Binary
We bypass one free_energy computation, making this about 25% faster.
The idea is that in the matrix E_ where we are computing the
free energies of all corrupted variants, we also compute in passing
the free energy of the original data.
So instead of calling `free_energy`, we extract the energy of the
original data from `E_`, and then take the difference.

The same idea can be extended to the other substitution_matrix
methods. Will do later.